+++ R/time_diff_gcd.R
@@ -15,9 +15,6 @@
 #' @param time_type If "auto", `periods` are used if `x` is a Date and
 #' durations are used if `x` is a datetime.
 #' Otherwise numeric differences are calculated.
-#' @param is_sorted Is `x` sorted? If `FALSE`, the default,
-#' `sort(unique(x))` is calculated, otherwise `unique(x)` is used.
-#' The order of successive time differences may impact the result.
 #' @param tol Tolerance of comparison. The time differences are rounded
 #' using `digits = ceiling(abs(log10(tol)))` to try and avoid
 #' precision issues.
@@ -28,6 +25,7 @@
 #' @examples
 #' library(timeplyr)
 #' library(lubridate)
+#' library(cppdoubles)
 #' \dontshow{
 #' .n_dt_threads <- data.table::getDTthreads()
 #' .n_collapse_threads <- collapse::get_collapse()$nthreads
@@ -46,7 +44,7 @@
 #' time_diff_gcd(quarter_seq, time_by = "months", time_type = "duration")
 #'
 #' # Detects monthly granularity
-#' dplyr::near(time_diff_gcd(as.vector(time(AirPassengers))), 1/12, tol = 1e-06)
+#' double_equal(time_diff_gcd(as.vector(time(AirPassengers))), 1/12)
 #' \dontshow{
 #' data.table::setDTthreads(threads = .n_dt_threads)
 #' collapse::set_collapse(nthreads = .n_collapse_threads)
@@ -54,18 +52,10 @@
 #' @export
 time_diff_gcd <- function(x, time_by = 1,
                           time_type = getOption("timeplyr.time_type", "auto"),
-                          is_sorted = FALSE,
                           tol = sqrt(.Machine$double.eps)){
-  if (!is_sorted){
-    is_sorted <- is_sorted(x)
-  }
-  x <- collapse::funique(x, sort = !is_sorted)
+  x <- collapse::funique(x, sort = FALSE)
   if (length(x) == 1L && is.na(x)){
     return(NA_real_)
-  }
-  x <- collapse::na_rm(x)
-  if (length(x) == 0L){
-    return(numeric())
   }
   if (length(x) == 1L){
     return(1)
@@ -75,16 +65,52 @@ time_diff_gcd <- function(x, time_by = 1,
                         time_type = time_type,
                         g = NULL,
                         na_skip = FALSE)
-  tdiff <- collapse::fdiff.default(tdiff, fill = Inf, n = 1)
+  tdiff <- cpp_roll_diff(tdiff, k = 1L, fill = 0)
   log10_tol <- ceiling(abs(log10(tol)))
-  tdiff <- collapse::funique.default(
-    round(
-      abs(tdiff), digits = log10_tol
-    )
-  )
-  tdiff <- tdiff[cpp_which(double_gt(tdiff, 0, tol = tol))]
-  if (length(tdiff) == 1 && tdiff == Inf){
-    return(10^(-log10_tol))
-  }
-  collapse::vgcd(tdiff)
+  tdiff <- round(abs(tdiff), digits = log10_tol + 1)
+  tdiff <- collapse::funique.default(tdiff)
+  gcd(tdiff, tol = tol, na_rm = TRUE, round = FALSE)
+  # cpp_gcd(as.double(tdiff), tol = as.double(tol),
+  #         start = 1L,
+  #         break_early = TRUE,
+  #         na_rm = TRUE,
+  #         round = FALSE)
 }
+
+# Previous method
+# time_diff_gcd2 <- function(x, time_by = 1,
+#                           time_type = getOption("timeplyr.time_type", "auto"),
+#                           is_sorted = FALSE,
+#                           tol = sqrt(.Machine$double.eps)){
+#   x <- collapse::funique(x, sort = FALSE)
+#   if (!is_sorted && !is_sorted(x)){
+#     x <- sort(x, na.last = TRUE)
+#   }
+#   if (length(x) == 1L && is.na(x)){
+#     return(NA_real_)
+#   }
+#   x <- collapse::na_rm(x)
+#   if (length(x) == 0L){
+#     return(numeric())
+#   }
+#   if (length(x) == 1L){
+#     return(1)
+#   }
+#   tdiff <- time_elapsed(x, rolling = FALSE,
+#                         time_by = time_by,
+#                         time_type = time_type,
+#                         g = NULL,
+#                         na_skip = FALSE)
+#   tdiff <- cpp_roll_diff(tdiff, k = 1L, fill = Inf)
+#   log10_tol <- ceiling(abs(log10(tol)))
+#   tdiff <- collapse::funique.default(
+#     round(
+#       abs(tdiff), digits = log10_tol
+#     )
+#   )
+#   tdiff <- tdiff[cpp_which(double_gt(tdiff, 0, tol = tol))]
+#   if (length(tdiff) == 1 && tdiff == Inf){
+#     return(10^(-log10_tol))
+#   }
+#   collapse::vgcd(tdiff)
+# }